Split main.go into multiple files
