+++ go.mod
@@ -3,6 +3,7 @@ module reporter
 go 1.15
 
 require (
+	github.com/aws/aws-lambda-go v1.22.0
 	github.com/aws/aws-sdk-go v1.36.25
 	github.com/stretchr/testify v1.7.0
 )
